+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## 69.0.0-SNAPSHOT - unreleased
+## 68.1.0 - 2024-09-27
 
 ### 🎁 New Features